Extract ConfigDatabaseGetCommand
